# service/openapi.yaml
openapi: 3.0.0
info:
  title: Job API
  description: API for managing job listings.
  version: "1.0"
paths:
  /jobs:
    get:
      summary: List all jobs
      responses:
        "200":
          description: A list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Job"
      responses:
        "201":
          description: The job was successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      summary: Get a single job by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the job
      responses:
        "200":
          description: Details of a job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /add-job:
    post:
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Job"
      responses:
        "201":
          description: The job was successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
components:
  schemas:
    Job:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        location:
          type: string
        job_description:
          type: string
        salary:
          type: number
        company_name:
          type: string
        company_description:
          type: string
        company_email:
          type: string
          format: email
        company_contact_number:
          type: string